Rename 'api_types' to 'models'
